+++ crates/steamlens-app/src/routing.rs
@@ -102,7 +102,7 @@ pub(crate) fn open_game_view(app: &mut App, app_id: u32) -> Task<Message> {
 
     if steam_off {
         if state.achievements.is_empty() && !app.context.cached_entries.contains_key(&app_id) {
-            let steamid3 = app.context.steamid3 as u32;
+            let steamid3 = app.context.steamid3;
             tasks.push(Task::perform(
                 cache::store::load_game_cache(steamid3, app_id),
                 move |entry| {
